Cgo: eliminate partial link step, update some providers
Previously cgo compiled generated .c files, along with the other C
sources in a package, then partially linked them all into a single .o
file which was packed into the .a file for the library. This partial
linking caused problems on platforms where -pie is on by default
(newer Debian and Ubuntu) because the -r flag we used conflicted.

The partial link is not necessary and is being removed from
"go build", so we will remove it, too. The .a file generated by the
cgo cc_library rule is unpacked, and its contents are packed into the
.a file for the go_library.

Fixes #252