Add orig_srcs field to GoSource and GoArchiveData
srcs carries the set of sources to be passed to the compiler. These
sources may be copied or transformed by cgo and cover and may have
files missing because they are passed to a non-Go compiler.

orig_srcs carries the set of source files passed into go_library or
compatible rules.