go_sdk: replace globs with generated file lists
When Bazel builds a target that depends on a glob, it traverses a
directory tree, making a list of matching files. This needs to be done
on every build to confirm files haven't been added or removed.

In repository rules, there's no way for files to be added or removed,
so globs are pointless. Instead, we can generate file lists ahead of
time and depend on those.

Fixes #1862