Fix go_googleapis directives patch for alpine
The first patch for go_googleapis deletes all build files, except the
root build file, which is empty. diff -urN doesn't generate a hunk for
this.

The second patch expects the root build file to be absent. The version
of patch that alpine uses reports an error because the file is still
present.

This PR fixes the timestamp on the second patch so the file is
expected to be present but empty.

Fixes #2042